// app/src/types/product.ts
export interface ProductData {
  id: number
  productCode: string
  name: string
  season: string
  color: string
  series: string
  fitType: string
  collarType: string
  shoulderType: string
  silhouette: string
  hemType: string
  fabricComposition: string
  fabricDesc: string
  craftSellingPoint: string
  fabricCareNotes: string
  suitableScenes: string
  imageUrl: string
}